更新了雷达航迹测试、火控解算精度测试、反应时间测试及发射架调转精度测试算法

+++ modules/weapon-test-api/src/main/java/com/soul/weapon/entity/HistoryInfo.java
@@ -422,10 +422,10 @@ public class HistoryInfo implements Serializable {
         private String weaponType;
 
         @Display("发射架方位角调转精度")
-        private Long launcherPitchAccuracy;
+        private Float launcherPitchAccuracy;
 
         @Display("发射架俯仰角调转精度")
-        private Long launcherAzimuthAccuracy;
+        private Float launcherAzimuthAccuracy;
 
     }
